+++ elfo-pinger/src/lib.rs
@@ -1,11 +1,11 @@
 #![warn(rust_2018_idioms, unreachable_pub)]
 
-use elfo_core::{ActorGroup, Schema, Topology};
+use elfo_core::{ActorGroup, Blueprint, Topology};
 
 mod actor;
 mod config;
 
-pub fn new(topology: &Topology) -> Schema {
+pub fn new(topology: &Topology) -> Blueprint {
     let topology = topology.clone();
     ActorGroup::new()
         .config::<config::Config>()